+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - [x] File icon for .editorconfig files
 - [x] Formatting command
 - [x] Validate numeric values
-- [ ] Quick Info tooltips
+- [x] Quick Info tooltips
 - [ ] Implement custom parser
 
 Features that have a checkmark are complete and available for